Noen justeringer
