+++ src/main/kotlin/no/nav/bidrag/behandling/felles/dto/stonad/MottakerIdHistorikkClient.kt
@@ -1,0 +1,5 @@
+package no.nav.bidrag.behandling.felles.dto.stonad
+
+interface MottakerIdHistorikkClient {
+  fun finnAlleEndringerMottakerIdForStonad(stonadId: Int): StonadDto
+}
