+++ src/main/kotlin/no/nav/bidrag/behandling/felles/dto/stonad/StonadClient.kt
@@ -1,0 +1,7 @@
+package no.nav.bidrag.behandling.felles.dto.stonad
+
+interface StonadClient {
+  fun hentStonad(stonadId: Int): StonadDto
+
+  fun endreMottakerIdOgOpprettHistorikk(request: EndreMottakerIdRequestDto): EndreMottakerIdDto
+}
